Fix infinite loop in VerifyTargetDimms
VerifyTargetDimms is used by many API entry points. This fix resolves
the issue of infinite loop that could result in the API not returning.

Loop counters type (UINT16) was incompatible with loop condition
parameter (UINT32) and caused the counters to wrap around.

+++ DcpmPkg/driver/Protocol/Driver/NvmDimmConfig.c
@@ -1341,8 +1341,8 @@ VerifyTargetDimms (
   LIST_ENTRY *pCurrentDimmNode = NULL;
   LIST_ENTRY *pDimmList = NULL;
   DIMM *pCurrentDimm = NULL;
-  UINT16 Index = 0;
-  UINT16 Index2 = 0;
+  UINT32 Index = 0;
+  UINT32 Index2 = 0;
   BOOLEAN FoundMatchDimmId = FALSE;
   BOOLEAN FoundMatchSocketId = FALSE;
